requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
openai==1.35.0
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1

